Исправляет замечания
