+++ src/components/not-found-screen/not-found-screen.jsx
@@ -1,15 +1,7 @@
 import React from 'react';
 import {Link} from 'react-router-dom';
-import {connect} from "react-redux";
-import PropTypes from "prop-types";
-import {logout} from "../../store/api-actions";
-import {AuthorizationStatus} from '../../const';
 
-const NotFoundScreen = ({authorizationStatus, onLogout}) => {
-  if (authorizationStatus === AuthorizationStatus.AUTH) {
-    onLogout();
-  }
-
+const NotFoundScreen = () => {
   return (
     <div className="page">
       <section className="page__main">
@@ -20,20 +12,4 @@ const NotFoundScreen = ({authorizationStatus, onLogout}) => {
   );
 };
 
-NotFoundScreen.propTypes = {
-  onLogout: PropTypes.func.isRequired,
-  authorizationStatus: PropTypes.string.isRequired,
-};
-
-const mapStateToProps = (state) => ({
-  authorizationStatus: state.authorizationStatus,
-});
-
-const mapDispatchToProps = (dispatch) => ({
-  onLogout() {
-    dispatch(logout());
-  },
-});
-
-export {NotFoundScreen};
-export default connect(mapStateToProps, mapDispatchToProps)(NotFoundScreen);
+export default NotFoundScreen;
