+++ src/store/action.js
@@ -5,7 +5,7 @@ export const ActionType = {
   REQUIRED_AUTHORIZATION: `user/requiredAuthorization`,
   LOAD_HOTELS: `data/loadHotels`,
   REDIRECT_TO_ROUTE: `login/redirectToRoute`,
-  GET_AUTH_INFO: `data/authInfo`,
+  SET_AUTH_INFO: `data/authInfo`,
 };
 
 export const ActionCreator = {
@@ -32,8 +32,8 @@ export const ActionCreator = {
     type: ActionType.REDIRECT_TO_ROUTE,
     payload: url,
   }),
-  getAuthInfo: (auth) => ({
-    type: ActionType.GET_AUTH_INFO,
+  setAuthInfo: (auth) => ({
+    type: ActionType.SET_AUTH_INFO,
     payload: auth,
   })
 };
